RF: Refractoring the function to be reused.
