+++ fury/actor/tests/_helpers.py
@@ -6,98 +6,121 @@ import numpy.testing as npt
 
 from fury import actor, window
 
+# --- Shared real inputs for visibility/snapshot tests -----------------------
 
-def _to_255(colors):
-    """
-    Convert one or more ``[0, 1]`` RGB colors to integer ``[0, 255]`` tuples.
-
-    Parameters
-    ----------
-    colors : sequence
-        A single ``(r, g, b)`` color or a sequence of them, with components in
-        ``[0, 1]``.
-
-    Returns
-    -------
-    list of tuple
-        The colors as integer RGB tuples suitable for
-        :func:`fury.window.analyze_snapshot`.
-    """
-    arr = np.atleast_2d(np.asarray(colors, dtype=float))
-    return [tuple(int(round(c * 255)) for c in row[:3]) for row in arr]
+CENTERS = np.array([[0.0, 0.0, 0.0]], dtype=np.float32)
+LINES = [np.array([[0, 0, 0], [1, 1, 1], [2, 0, 0]], dtype=np.float32)]
 
 
-def snapshot_report(scene, *, colors=None):
-    """
-    Render ``scene`` offscreen and analyze the resulting image.
-
-    A thin wrapper around :func:`fury.window.snapshot` (with
-    ``return_array=True``) and :func:`fury.window.analyze_snapshot`. Performs a
-    real offscreen render -- no mocks, patches, or dummy objects.
-
-    Parameters
-    ----------
-    scene : Scene
-        The scene to render.
-    colors : sequence of (r, g, b), optional
-        Expected colors in ``[0, 1]``. When given, the returned report's
-        ``colors_found`` reports whether each color is present in the image.
-
-    Returns
-    -------
-    ReportSnapshot
-        The analysis report, exposing ``objects`` (count of foreground blobs)
-        and ``colors_found``.
-    """
-    rgb255 = _to_255(colors) if colors is not None else None
-    arr = window.snapshot(scene=scene, fname=None, return_array=True)
-    return window.analyze_snapshot(arr, colors=rgb255, find_objects=True)
+def gradient_image(n=64):
+    """A non-uniform 2D image so the rendered frame is never flat."""
+    ramp = np.linspace(0.0, 1.0, n, dtype=np.float32)
+    return np.outer(ramp, ramp)
 
 
-def assert_visibility(target, *, toggle, colors=None):
-    """
-    Assert ``target`` renders when visible and renders nothing when hidden.
+def scalar_volume(n=32):
+    """A non-uniform 3D scalar volume (deterministic)."""
+    x, y, z = np.mgrid[0:n, 0:n, 0:n]
+    return ((x + y + z) % 17).astype(np.float32)
 
-    Drives a real offscreen render through :func:`snapshot_report` and uses the
-    background-difference object count from
-    :func:`fury.window.analyze_snapshot` as the visibility signal. No mocks,
-    patches, or dummy classes are involved -- visibility is toggled through the
-    object's real API via ``toggle``.
 
-    Parameters
-    ----------
-    target : Object
-        An actor or UI element. It is added to a fresh
-        :class:`~fury.window.Scene` for each render (``window.snapshot`` attaches
-        a camera to the scene, so a new scene is used per snapshot).
-    toggle : callable
-        The object's real visibility setter, invoked as ``toggle(True)`` and
-        ``toggle(False)`` (e.g. ``element.set_visibility`` for UI, or
-        ``lambda v: setattr(actor, "visible", v)`` for a raw actor).
-    colors : sequence of (r, g, b), optional
-        Expected colors in ``[0, 1]``. When given, they must be present while
-        visible and absent while hidden. Reliable only for flat/``"basic"``
-        materials where the rendered color matches exactly.
-    """
+def roi_volume(n=20):
+    """A binary volume with a solid central cube (for contour actors)."""
+    data = np.zeros((n, n, n), dtype=np.float32)
+    data[5:15, 5:15, 5:15] = 1.0
+    return data
 
-    def _render():
-        scene = window.Scene()
-        scene.add(target)
-        return snapshot_report(scene, colors=colors)
 
-    toggle(True)
-    report = _render()
-    assert report.objects >= 1, "expected a visible object to render"
-    if colors is not None:
-        assert all(report.colors_found), "expected colors missing while visible"
+def label_volume(n=20):
+    """A labeled volume with one nonzero region."""
+    data = np.zeros((n, n, n), dtype=int)
+    data[5:15, 5:15, 5:15] = 1
+    return data
 
-    toggle(False)
-    report = _render()
-    assert report.objects == 0, "expected nothing to render while hidden"
-    if colors is not None:
-        assert not any(report.colors_found), "colors present while hidden"
 
-    toggle(True)
+def peak_dirs(n=8):
+    """A (X, Y, Z, 3, 3) field of unit peak directions."""
+    dirs = np.zeros((n, n, n, 3, 3), dtype=np.float32)
+    dirs[..., 0, :] = (1.0, 0.0, 0.0)
+    dirs[..., 1, :] = (0.0, 1.0, 0.0)
+    dirs[..., 2, :] = (0.0, 0.0, 1.0)
+    return dirs
+
+
+def vector_field_data(n=8):
+    """A (X, Y, Z, 3) field of unit vectors."""
+    field = np.zeros((n, n, n, 3), dtype=np.float32)
+    field[...] = (1.0, 0.0, 0.0)
+    return field
+
+
+def sph_coeffs(n=4, ncoeff=15):
+    """SH coefficients with only the l=0 term -> isotropic visible glyphs."""
+    coeffs = np.zeros((n, n, n, ncoeff), dtype=np.float32)
+    coeffs[..., 0] = 1.0
+    return coeffs
+
+
+def triangle_mesh():
+    """A single-triangle (vertices, faces) pair for surface actors."""
+    verts = np.array([[0, 0, 0], [1, 0, 0], [0, 1, 0]], dtype=np.float32)
+    faces = np.array([[0, 1, 2]], dtype=np.int32)
+    return verts, faces
+
+
+# The full set of renderable actors keyed by id. Each value is a factory that
+# builds a fresh actor (with real inputs) on every call.
+ACTOR_FACTORIES = {
+    # Planar / polyhedron / curved mesh primitives (centers only).
+    "square": lambda: actor.square(CENTERS),
+    "triangle": lambda: actor.triangle(CENTERS),
+    "star": lambda: actor.star(CENTERS),
+    "disk": lambda: actor.disk(CENTERS),
+    "ring": lambda: actor.ring(CENTERS),
+    "box": lambda: actor.box(CENTERS),
+    "tetrahedron": lambda: actor.tetrahedron(CENTERS),
+    "icosahedron": lambda: actor.icosahedron(CENTERS),
+    "triangularprism": lambda: actor.triangularprism(CENTERS),
+    "pentagonalprism": lambda: actor.pentagonalprism(CENTERS),
+    "octagonalprism": lambda: actor.octagonalprism(CENTERS),
+    "rhombicuboctahedron": lambda: actor.rhombicuboctahedron(CENTERS),
+    "frustum": lambda: actor.frustum(CENTERS),
+    "superquadric": lambda: actor.superquadric(CENTERS),
+    "cylinder": lambda: actor.cylinder(CENTERS),
+    "cone": lambda: actor.cone(CENTERS),
+    "arrow": lambda: actor.arrow(CENTERS),
+    "sphere": lambda: actor.sphere(CENTERS),
+    "ellipsoid": lambda: actor.ellipsoid(CENTERS),
+    # Points / billboards.
+    "point": lambda: actor.point(CENTERS),
+    "marker": lambda: actor.marker(CENTERS),
+    "billboard": lambda: actor.billboard(CENTERS),
+    "billboard_sphere": lambda: actor.billboard_sphere(CENTERS),
+    # Axes (no required args).
+    "axes": lambda: actor.axes(),
+    # Text / image.
+    "text": lambda: actor.text("FURY"),
+    "image": lambda: actor.image(gradient_image()),
+    # Line / streamline family.
+    "line": lambda: actor.line(LINES),
+    "streamlines": lambda: actor.streamlines(LINES),
+    "streamtube": lambda: actor.streamtube(LINES),
+    "line_projection": lambda: actor.line_projection(LINES),
+    # Surface.
+    "surface": lambda: actor.surface(*triangle_mesh()),
+    # Volume / field / glyph slicers (return single actors or groups; both are
+    # hidden by setting ``.visible`` on the returned object).
+    "data_slicer": lambda: actor.data_slicer(scalar_volume()),
+    "volume_slicer": lambda: actor.volume_slicer(scalar_volume()),
+    "peaks_slicer": lambda: actor.peaks_slicer(peak_dirs()),
+    "vector_field": lambda: actor.vector_field(vector_field_data()),
+    "vector_field_slicer": lambda: actor.vector_field_slicer(vector_field_data()),
+    "sph_glyph": lambda: actor.sph_glyph(sph_coeffs()),
+    # Contours (need a solid block to produce geometry).
+    "contour_from_volume": lambda: actor.contour_from_volume(roi_volume()),
+    "contour_from_roi": lambda: actor.contour_from_roi(roi_volume()),
+    "contour_from_label": lambda: actor.contour_from_label(label_volume()),
+}
 
 
 def random_png(width, height):
