+++ fury/testing.py
@@ -47,6 +47,93 @@ def captured_output():
         sys.stdout, sys.stderr = old_out, old_err
 
 
+def snapshot_report(scene, *, colors=None):
+    """
+    Render ``scene`` offscreen and analyze the resulting image.
+
+    A thin wrapper around :func:`fury.window.snapshot` (with
+    ``return_array=True``, writing no file) and
+    :func:`fury.window.analyze_snapshot`. Performs a real offscreen render --
+    no mocks, patches, or dummy objects.
+
+    Parameters
+    ----------
+    scene : Scene
+        The scene to render.
+    colors : sequence of (r, g, b), optional
+        Expected colors in ``[0, 1]``. When given, the returned report's
+        ``colors_found`` reports whether each color is present in the image.
+
+    Returns
+    -------
+    ReportSnapshot
+        The analysis report, exposing ``objects`` (count of foreground blobs)
+        and ``colors_found``.
+    """
+    from fury.colormap import colors_to_uint8
+    from fury.window import analyze_snapshot, snapshot
+
+    rgb255 = colors_to_uint8(colors) if colors is not None else None
+    arr = snapshot(scene=scene, fname=None, return_array=True)
+    return analyze_snapshot(arr, colors=rgb255, find_objects=True)
+
+
+def assert_visibility(target, *, toggle, colors=None):
+    """
+    Assert ``target`` renders when visible and renders nothing when hidden.
+
+    Drives a real offscreen render through :func:`snapshot_report` and uses the
+    background-difference object count from
+    :func:`fury.window.analyze_snapshot` as the visibility signal. No mocks,
+    patches, or dummy classes are involved -- visibility is toggled through the
+    object's real API via ``toggle``.
+
+    Parameters
+    ----------
+    target : Object
+        An actor or UI element. It is added to a fresh
+        :class:`~fury.window.Scene` for each render (``window.snapshot`` attaches
+        a camera to the scene, so a new scene is used per snapshot).
+    toggle : callable
+        The object's real visibility setter, invoked as ``toggle(True)`` and
+        ``toggle(False)`` (e.g. ``element.set_visibility`` for UI, or
+        ``lambda v: setattr(actor, "visible", v)`` for a raw actor).
+    colors : sequence of (r, g, b), optional
+        Expected colors in ``[0, 1]``. When given, they must be present while
+        visible and absent while hidden. Reliable only for flat/``"basic"``
+        materials where the rendered color matches exactly.
+    """
+    from fury.window import Scene
+
+    def _render():
+        """
+        Render function to check visibility.
+
+        Returns
+        -------
+        ReportSnapshot
+            The analysis report, exposing ``objects`` (count of foreground blobs)
+            and ``colors_found``.
+        """
+        scene = Scene()
+        scene.add(target)
+        return snapshot_report(scene, colors=colors)
+
+    toggle(True)
+    report = _render()
+    assert report.objects >= 1, "expected a visible object to render"
+    if colors is not None:
+        assert all(report.colors_found), "expected colors missing while visible"
+
+    toggle(False)
+    report = _render()
+    assert report.objects == 0, "expected nothing to render while hidden"
+    if colors is not None:
+        assert not any(report.colors_found), "colors present while hidden"
+
+    toggle(True)
+
+
 def assert_operator(value1, value2, *, msg="", op=operator.eq):
     """
     Check boolean statement using the given operator.
